Move traits into a submodule
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
