+++ src/lib.rs
@@ -1,12 +1,3 @@
-mod broadcast;
-mod contains;
-mod mask;
-mod merge;
-pub use broadcast::*;
-pub use contains::*;
-pub use mask::*;
-pub use merge::*;
-
 mod netaddr;
 mod netv4addr;
 mod netv6addr;
@@ -16,6 +7,9 @@ pub use netv6addr::*;
 
 mod netaddr_error;
 pub use netaddr_error::*;
+
+mod traits;
+pub use traits::*;
 
 #[cfg(test)]
 mod tests;
